Realocado dict correspodencia em seus respectivos aplicativos
