--- article/exceptions/article_exceptions.py
@@ -1,2 +1,0 @@
-class ArticleSaveError(Exception):
-    ...
