+++ issue/utils/correspondencia.py
@@ -1,0 +1,23 @@
+correspondencia_issue = {
+    "v31": "volume",
+    "v32": "number",
+    'v33': 'issue_title',
+    'v34': 'part',
+    'v35': 'scielo_issn',
+    'v42': 'status',
+    "v43": "bibliographic_strip",
+    'v48': 'header_of_table_of_contents',
+    'v49': 'sections_data',
+    'v62': 'issue_editor',
+    'v65': 'date_iso',
+    'v85': 'controlled_vocabulary',
+    'v97': 'cover',
+    'v117': 'standard',
+    'v122': 'number_of_documents',
+    "v131": "supllement",
+    "v132": "supllement",
+    'v140': 'sponsor',
+    'v200': 'markup_done',
+    'v435': 'issn',
+    'v540': 'identifying_regular_issues'
+}
